#30 split tableau into separate file
